Added marmot.Agent inheritance to Vessel. Moved Vessel and components to ORBIT.phases.install._core
