+++ ORBIT/phases/install/monopile_install/standard.py
@@ -8,10 +8,10 @@ __email__ = "jake.nunemaker@nrel.gov"
 
 import simpy
 
-from ORBIT.vessels import Vessel
 from ORBIT.simulation import Environment, VesselStorage
 from ORBIT.phases.install import InstallPhase
 from ORBIT.simulation.logic import shuttle_items_to_queue
+from ORBIT.phases.install._core import Vessel
 from ORBIT.phases.install.monopile_install._single_wtiv import (
     solo_install_monopiles,
 )
