+++ ORBIT/phases/install/turbine_install/standard.py
@@ -12,10 +12,10 @@ from math import ceil
 import numpy as np
 import simpy
 
-from ORBIT.vessels import Vessel
 from ORBIT.simulation import Environment, VesselStorage
 from ORBIT.phases.install import InstallPhase
 from ORBIT.simulation.logic import shuttle_items_to_queue
+from ORBIT.phases.install._core import Vessel
 from ORBIT.phases.install.turbine_install._single_wtiv import (
     solo_install_turbines,
 )
