+++ ORBIT/vessels/__init__.py
@@ -4,5 +4,3 @@ __author__ = ["Jake Nunemaker", "Rob Hammond"]
 __copyright__ = "Copyright 2020, National Renewable Energy Laboratory"
 __maintainer__ = "Jake Nunemaker"
 __email__ = "jake.nunemaker@nrel.gov"
-
-from .base import Vessel
