--- ORBIT/vessels/components/__init__.py
@@ -1,14 +1,0 @@
-"""
-This subpackage contains classes and methods that describe optional vessel
-components.
-"""
-
-__author__ = ["Jake Nunemaker", "Rob Hammond"]
-__copyright__ = "Copyright 2020, National Renewable Energy Laboratory"
-__maintainer__ = "Jake Nunemaker"
-__email__ = "jake.nunemaker@nrel.gov"
-
-
-from .crane import Crane
-from .jacksys import JackingSys
-from .carousel import Carousel, CarouselSystem
